Update Single Response

+++ src/LMS.Domain/IRepositories/IServerRepository.cs
@@ -4,12 +4,11 @@ namespace BigBlueApi.Domain.IRepository;
 
 public interface IServerRepository
 {
-    ValueTask<bool> CanJoinServer(int id);
-    ValueTask<int> DeleteServer(int Id);
-    ValueTask<Server> GetServer(int Id);
+    ValueTask<bool> CanJoinServer(long id);
+    ValueTask<long> DeleteServer(long Id);
+    ValueTask<Server> GetServer(long Id);
     ValueTask<List<Server>> GetAllServer();
     ValueTask<Server> CreateServer(Server server);
     ValueTask<Server> MostCapableServer();
-
-    Task EditServer(int id, Server server);
+    Task EditServer(long id, Server server);
 }
